Bug 3585: Shift key appears as rectangle instead of up arrow
Root Cause: The "Liberation" fonts that are installed with the base yocto image
does not contain a glyph for the "\u21E7" codepoint, so it displays a
Placeholder image instead of the shift key.

Solution: Install the "DejaVu" font that supports the codepoint.

Why and When Introduced: Initial commit. Fonts were not considered.

Possible Similar Problems: Checked the alphanumeric keyboard and numpad for any
other unicode points that aren't supported.

Testing: Run bitbake micromeritics-1350-image and check for the new font files.
Install the font files on the i1350 prototype and check the appearance of the
shift key.

Recommendation: We may want to consider using fonts from Blackhagen, and we will
need to add more fonts when we support more languages.