Bug 3550: Scrolling in the instrument touch screen does not work as expected.
Root Cause:
Chromium was not configured to run with the proper flags.

Solution:
The biggest part of this solution was figuring out how to get the
argument of --touch-devices. The argument of this flag expects an
ID reported by xinput, which represents the touchscreen device. It was
discovered that the ID reported by xinput is not consistent over all the
touchscreens (from different manufacturers) being used.
As a result, each xinput device has to be checked with udev to see if
the input device has the following two parameters:
  1. ID_BUS=usb
  2. ID_INPUT_TOUCHSCREEN=1

Testing:
Ran the code on the SBC and verified that scrolling works by touching
any part of the screen and moving up or down. The scrollbar also only
appears as long as there's contact.