Bug 3540: Disable CTRL functions on connected keyboard
Root Cause:
Users are able to use keyboard shortcuts to enter a linux terminal or to
open/close chromium tabs. This looks bad and should not be allowed.

Solution:
Disable the Ctrl, Alt, and F[1-12] keys.

The SBC uses xserver-nodm as its display-manager. This is started as a
systemd unit from /etc/systemd/system/display-manager.service, which runs the
/etc/xserver-nodm/Xserver shell script.
The Xserver script starts the xinit binary with the following script:
/etc/X11/Xsession
The Xsession script runs all the shell scripts (in ascending alphabetical order) found in:
/etc/X11/Xsession.d/*.sh
Before this commit, the files in that directory were:
13xdgbasedirs.sh  30xinput_calibrate.sh  89xdgautostart.sh  90XWindowManager.sh

80xmodmap.sh was added in the above directory so that it disables the
relevant keys before the chromium session is started.
Note: xev had to be temporarily bitbaked into the image in order to find
      out the X keycodes corresponding to keyboard strokes. This was
      done by temporarily adding the following to recipes-local/images/micromeritics-basesystem-image.bb:
          IMAGE_INSTALL += " xev "

Testing:
Ran the bitbake and tested the image to verify that F1 did not open a
new browser window, and that Ctrl + t/w did not open/close new tabs on
chrome.