Fixed scrolling and detect keyboard/mouse attach/detach events.
Solution:
Fixed touchscreen identification by using xinput to detect the device
node that responds to the user's touch.

Added pyusb to detect USB attach/detach events. Replaced unclutter by passing -nocursor
flag to the display manager.

A systemd unit is spawned to run a python script
(micromeritics-kb-mouse-detection.py) which polls for USB device
changes. If a keyboard or mouse is detected (by the product name
containing either "keyboard" or "mouse"), then:
1. The xorg file is updated with the -nocursor argument.
2. The mic_python_lib.Settings table has two members - has_phys_kb,
   has_phys_mouse, which are updated appropriately.
   Direct SQL operations are used instead of going through django to
   access the database in order to improve startup time by 2 seconds.
   Also, changed systemd unit type from simple to notify, in order to
   make sure that the /etc/default/xserver-nodm file is correct, before
   starting the xserver-nodm systemd unit.

Testing:
Connect a mouse, restart the SBC, and note the presence of a cursor.
Disconnect the mouse, restart the SBC, and note the absence of a cursor.

Two mice and one keyboard were used for these tests.
The database can be checked by running the following on the SBC:
> python3 /usr/bin/manage.py shell
> from mic_python_lib.models import Settings
> (Settings.objects.first().has_phys_kb, Settings.objects.first().has_phys_mouse)
Bitbake the image and transfer the image onto the SBC.

Connect the keyboard. Verify that the table's has_phys_kb is True and
has_phys_mouse is False.

Connect the mouse. Verify that the /etc/default/xserver-nodm file is
updated with the -nocursor string. Verify that the table's has_phys_kb is True and
has_phys_mouse is True.

Connect the second mouse. Verify that the /etc/default/xserver-nodm file is
updated with the -nocursor string. Verify that the table's has_phys_kb is True and
has_phys_mouse is True.

Disconnect the second mouse. Verify that the /etc/default/xserver-nodm file still
contains the -nocursor string. Verify that the table's has_phys_kb is True and
has_phys_mouse is True.

Disconnect the keyboard. Verify that the /etc/default/xserver-nodm file still
contains the -nocursor string. Verify that the table's has_phys_kb is False and
has_phys_mouse is True.

Disconnect the mouse. Verify that the /etc/default/xserver-nodm file
does not contain the -nocursor string. Verify that the table's has_phys_kb is False and
has_phys_mouse is False.