Added/Removed necessary packages for the host and target in the SDK.
Testing:
Run the following command in the docker-yocto-build container.
  bitbake micromeritics-basesystem-image -c populate_sdk
Tar the SDK from the build computer and transfer it to the development
maachine:
  tar cvhzf ~/sdk.tar.gz docker-yocto-build/1350_vol/yocto-sbc-iot-imx7/build-x11-imx7/tmp/deploy/sdk
Untar the above sdk.tar.gz file on the development machine.
  cd ~/ && tar -xvzf ~/sdk.tar.gz
Run the SDK-installer script:
  bash ~/sdk/fsl-imx-x11-glibc-x86_64-micromeritics-basesystem-image-cortexa7hf-neon-toolchain-4.19-warrior.sh
Souce the SDK:
  source /opt/fsl-imx-x11/4.19-warrior/environment-setup-cortexa7hf-neon-poky-linux-gnueabi
Build the i1350 for the target with the cross-compiler:
   cd ~/i1350 && cmake -DCMAKE_BUILD_TYPE=Debug -DCMAKE_VERBOSE_MAKEFILE:BOOL=ON -DBoost_DEBUG=1 -S . -B build-debug && cmake --build build-debug