Add ability to program instrument software from a usb stick on bootup.
Add the program-instrument python script as a service to the
bootblock.  This will check to see if an instrument has been
installed, and if not, wait a bit for a usb stick to be installed, and
if it finds instrument software on it, it will install it and reboot
with it installed.

Also:
1. Install a mic-basesystem file (from here) into /etc as an
   indication of the version of the basesytem used.  (just always
   1.0.0 for now.)
2. Make the versions of the mic-specific components at 1.0.0 to match
   the expected initial release version.