Bug 3612: Resolution should be native on asus sbc/screen. Also resets to 1024x768 after updating software.
Root cause:
The bitbaked linux image was generating a /etc/xdg/weston/weston.ini
file which set the output resolution to 1920x1080.

Solution:
The bitbake now uses a sed command in the do_install operation to modify
the /etc/xdg/weston/weston.ini file to have a resolution of 720x480.
This solution was used because bitbake was not picking up a changed ini
file in a "files"directory. Even cp or install commands to overwrite the
ini file were not working properly.

The template was taken from meta-asus-imx/recipes-graphics/wayland/weston-init