Bug 3536: Cursor appears when it should not.
Root cause:
This was originally fixed in 9cfed5dce4e1c08cabbc2995cc390163c5ee483d.
However, the board was changed from one that was using the X server to
one that was using wayland. As a result, the -nocursor flag no longer
works.

Solution:
Wayland makes use of the weston compositor which has a hide-cursor flag
in later versions. This patch was found on gitlab, downloaded, and
integrated into the weston build.
1. Get bitbake to modify the /etc/xdg/weston/weston.ini file by adding
   a "hide-cursor=false" default entry under the libinput section.
    recipes-graphics/wayland/weston-init.bbappend
2. Downloaded the "hide-cursor" patch from gitlab. The reason it isn't
   downloaded during the bitbake or build, is because this patch is currently
   on a non-integrated branch, which can change in the future.
    recipes-graphics/wayland/weston/0002-weston-implement-hide-cursor.patch
3. Add the patch to the source during the bitbake.
    recipes-graphics/wayland/weston_9.0.0.imx.bbappend

The USB detection script also had to change to modify the
/etc/xdg/weston/weston.ini file, appropriately.

Testing:
Added/removed mouse and noticed appropriate changes to /etc/xdg/weston/weston.ini.