Bug 3550: Scrolling in the instrument touch screen does not work as expected.
Root cause:
This was originally fixed in 1e6c0ca665f1c4e7c5ddfa2f44e08276ae4e53b6
However, the board was changed from one that was using the X server to
one that was using wayland. As a result, the previous solution which
relies on xinput and udevadm no longer works properly.

Solution:
XWayland is being used on the board which has some backwards
compatibility with legacy X applications like xinput.
However, there is no longer a way to map /dev/input/event[0-9] to the
xinput ID, which was the previous solution.
The only solution that seemed to work through trial and error, is to use
the ID associated with the xwayland-pointer device (not the xwayland-touch
device).

Removed unused file.
    recipes-local/micro/files/40xinput_identify_touchscreen.sh
    recipes-local/micro/micro-boot-chromium_1.0.0.bb