Bug 804: Mouse cursor does not appear when mouse is connected
It is not clear how well the previous service was working.  It had several things
incorrect, including (I think) starting before the the filesystem was available.  But it
ran continually (till the database lock caused it to throw and exception and exit) to
update this.   I believe this resulted in needing two reboots to get the weston
setting (the only one that was used) to be correct.

Update micromeritics-kb-mouse-detection to be a one shot service.

The spec (so far) for the instrument is that it only detects and responds to changes usb
mouse and keybard events when it is first turned on.  (The behavior can be somewhat funky,
but not disastrous if they are changed while running.)

So, this service now does not check continually, but only during bootup.  It also no
longer sets the database fields.  This did not work anyway because at least for some
times, the database was locked, and this caused an unhanded exception that ended the
service pretty quickly.

There is now an /etc/micromeritics-config.ini that holds the bootup mouse and keybard
discovery results.

Testing:
1. The script was run manually with different combinations of keybard and mouse plugged
   in, and verified to have created/deleted the indicator files.
2. Made a new basesystem and rebooted with and without mouse to see if the weston pointer
   was properly initialized.