Bug 808: Mouse scrolling does not work on some screens
Root Cause:  (From Cameron) The --touch-devices option tells chromium to interpret the
events from specified devices as touch events. This is what allows "push scrolling" with
your finger. When the mouse is connected, the mouse scroll events from the xserver are
interpreted as touch events, causing the behavior where scrolling with the cursor over an
element will click on that element.

Solution: Since we only support adding/removing mouse and keyboard while the power is off,
setup the mic-chromium script to check to see if a mouse was plugged in during boot (using
the fix for bug 804).  If it is, don't add the touch command line arguments, if it is not
keep them as they had been used.

Testing: Tested with both mouse plugged in and not plugged in (after re-running the
kb-mouse-detection service), made sure on restart that the mouse wheel did not trigger
touch events (when plugged in), and the touch scrolling worked when not plugged in.