Bug: MettlerToledo mass balance did not send mass to entry field.
Root Cause:  The mass balance sent keyboard sequence over the usb but this was interpreted
as control charaters.  (It worked in windows with excel/notepad.) That is because the
numlock was not on.

Solution: Add the numlockx package to the basesystem.  Then call this to set the numlock
to on as part of the mic-chromium service.  This is a good spot to load it since it is
here that the window server is running, and it is not needed except inside chromium.

Why and When Introduced:  Linux just disables this by default on startup.

Possible Similar problems: Where Looked, What Found, and What Action: NA

Testing: Made a clean basesystem build and verified that the MettlerToledo system will
work.  Also verified that the Sartorious balance still works.