Automatically start the pulseaudio and re-connect to know bluetooth speaker.
I did this as part of the mic-chromium service that runs as part of the framework.  This
is the only piece that needs this, so that is as good a place as any.  Note however that
this means that when the service is stopped the pulseaudio (and speaker connection) is
also blocked.