Bug: The 1350 was being reprogrammed every time it was booted if the usb had an application.
Root Cause:  The /etc/mic-instrument file was not present during boot.  That file
indicates it is already an instrument.   The reason that was not there is because the
program-instrument.py did not have a / between the usb directory and the instrument
directory when it tried to copy the file.

Solution: Properly build the filename to copy.

Why and When Introduced: Probably initial implementation (although testing should have
caught this).  Maybe the trailing slash on the usb dirs was removed at some point after
testing.

Possible Similar problems: Where Looked, What Found, and What Action: NA.

Testing: Made sure the /etc/mic-instrument file is installed on reboot.