Fix the logrotate that caused the instruemnt to be really slow after a time.
The logrotate was setup with wildcards for samba and nginx.  This caused the logrotate to
rotate the already rotated logs causing an explosion of backup files.   In particular the
10 cm3 unit in the apps lab was experiencing serious slow downs.  These were in part due
to the logrotate command using 100% of cpu for long periods of time while it was trying to
rotate too many files.  The /var/log/nginx and /var/log/samba both had several hundred
thousand files in them that were attempted to be rotated.

This makes sure we only rotate the specific files we want, and so don't have this
recursion.