RM 1251: Basesystem still uses up too much disk space for log.
Root Cause: The nginx config file tried to turn off logging by setting the error and
access logs to off.  But this really meant that the error and access logs went to a file
called off (for some reason) in the /usr/ directory.

Solution: Set the files to /dev/null in the usual unix way.

Why and When Introduced: The last time I tried to fix the logging.

Testing: Looked for the /usr/off file in a clean basystem and 1350 install before this
change(found), and then again after(not found).