RM 1473: Updated yocto basesystem that has the app level changes in it.
This adds the localhost => 127.0.0.1 change that was made for occasional failures in
forwarded http requests in the 1350 and 7100.  The nginx logging mistake was already done
in the RM 1251 commit.

Both of these were tested by (temporarily) removing the application level install fix, and
re-flashing the basesystem, installing the application, and seeing that the setup
files (/etc/nginx/sites-available/default_server and /etc/nginx/nginx.conf) are correct.
Then adding the application level install and installing the application to make sure
there are no issues.